+++ release_notes.md
@@ -7,3 +7,4 @@
  * Create the main application loop in the engine and create a application handle to manage client application lifetime.
  * Add a basic memory system into the engine that tracks memory usage by tagged allocation.
  * Added darray container API
+ * Added an immediate mode events system